Cast to double and fix normalize

+++ src/ortographicCamera.cpp
@@ -7,7 +7,7 @@ OrtographicCamera::OrtographicCamera(){}
 
 OrtographicCamera::OrtographicCamera(
     Point e, Vector3 u, Vector3 v, Vector3 w,
-    std::tuple<float, float, float, float> screenWindow
+    std::tuple<double, double, double, double> screenWindow
 ) : Camera(e, u, v, w, screenWindow) {}
 
 Ray OrtographicCamera::generate_ray(int x, int y)
